riscv: get configstring parsing to build
