+++ sys/src/libc/libc.json
@@ -151,6 +151,7 @@
 		"port/cistrstr.c",
 		"port/charstod.c",
 		"port/cleanname.c",
+		"port/configstring.c",
 		"port/ctype.c",
 		"port/encodefmt.c",
 		"port/execl.c",
